Add support for Laminas-db
