+++ composer.json
@@ -59,7 +59,7 @@
         "phpunit/phpunit": "^7.5 || ^8.5",
         "propel/propel1": "~1.7",
         "psr/log": "~1.0",
-        "zendframework/zend-db": "~2",
+        "laminas/laminas-db": "^2.0",
         "symfony/routing": "^4.4 || ^5.0",
         "symfony/http-kernel": "^4.4 || ^5.0",
         "symfony/http-foundation": "^4.4 || ^5.0",
