+++ src/Payum/Core/Tests/Bridge/Zend/Storage/TableGatewayStorageTest.php
@@ -2,8 +2,9 @@
 namespace Payum\Core\Tests\Bridge\Zend\Storage;
 
 use Payum\Core\Bridge\Zend\Storage\TableGatewayStorage;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Zend\Db\TableGateway\TableGateway;
+use Laminas\Db\TableGateway\TableGateway;
 
 class TableGatewayStorageTest extends TestCase
 {
@@ -38,10 +39,10 @@ class TableGatewayStorageTest extends TestCase
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|TableGateway
+     * @return MockObject|TableGateway
      */
     protected function createTableGatewayMock()
     {
-        return $this->createMock('Zend\Db\TableGateway\TableGateway', array(), array(), '', false);
+        return $this->createMock(TableGateway::class);
     }
 }
